Refactor ILoop
