+++ scio-repl/src/main/scala/com/spotify/scio/repl/ScioILoop.scala
@@ -19,14 +19,12 @@ package com.spotify.scio.repl
 
 import com.spotify.scio.BuildInfo
 import com.spotify.scio.bigquery.BigQuerySysProps
-
 import org.apache.beam.sdk.extensions.gcp.options.GcpOptions.DefaultProjectFactory
 import org.apache.beam.sdk.options.PipelineOptionsFactory
 import org.apache.commons.text.StringEscapeUtils
 
-import scala.tools.nsc.{CompilerCommand, Settings}
+import scala.tools.nsc.CompilerCommand
 import scala.tools.nsc.interpreter.Results
-import scala.tools.nsc.interpreter.shell.{ILoop, ShellConfig}
 
 /**
  * ScioILoop - core of Scio REPL.
@@ -34,7 +32,7 @@ import scala.tools.nsc.interpreter.shell.{ILoop, ShellConfig}
  * @param args user arguments for Scio REPL
  */
 class ScioILoop(command: CompilerCommand, scioClassLoader: ScioReplClassLoader, args: List[String])
-    extends ILoop(ShellConfig(command.settings)) {
+    extends compat.ILoop(command) {
 
   // Fail fast for illegal arguments
   try {
@@ -154,7 +152,7 @@ class ScioILoop(command: CompilerCommand, scioClassLoader: ScioReplClassLoader, 
   private val scioCommands = List(newScioCmd, newLocalScioCmd, scioOptsCmd)
 
   // TODO: find way to inject those into power commands. For now unused.
-  private val scioPowerCommands = List(createJarCmd, getNextJarCmd, runScioCmd) // scalafix:ok
+  private val scioPowerCommands = List(createJarCmd, getNextJarCmd, runScioCmd)
 
   override def commands: List[LoopCommand] = super.commands ++ scioCommands
 
@@ -171,7 +169,7 @@ class ScioILoop(command: CompilerCommand, scioClassLoader: ScioReplClassLoader, 
     s"""$factory.fromArgs($optionsAsStr).create()"""
   }
 
-  override def welcome(): String = {
+  override def welcome: String = {
     val p = scala.util.Properties
     val ascii =
       """Welcome to
@@ -236,15 +234,13 @@ class ScioILoop(command: CompilerCommand, scioClassLoader: ScioReplClassLoader, 
         |import _ioCommands._
       """.stripMargin)
 
-  override def createInterpreter(settings: Settings): Unit = {
-    super.createInterpreter(settings)
+  override def initCommand(): Unit = {
     intp.beQuietDuring {
       addImports()
       createBigQueryClient()
       newScioCmdImpl("sc")
       loadIoCommands()
     }
-    out.print(prompt)
-    out.flush()
   }
+
 }
